Add SurveyDiffCallback class
