+++ app/src/main/java/org/ole/planet/myplanet/ui/survey/AdapterSurvey.kt
@@ -264,18 +264,6 @@ class AdapterSurvey(
     }
 }
 
-class SurveyDiffCallback(private val oldList: List<RealmStepExam>, private val newList: List<RealmStepExam>) : DiffUtil.Callback() {
-    override fun getOldListSize(): Int = oldList.size
-    override fun getNewListSize(): Int = newList.size
-    override fun areItemsTheSame(oldItemPosition: Int, newItemPosition: Int): Boolean {
-        return oldList[oldItemPosition].id == newList[newItemPosition].id
-    }
-
-    override fun areContentsTheSame(oldItemPosition: Int, newItemPosition: Int): Boolean {
-        return oldList[oldItemPosition] == newList[newItemPosition]
-    }
-}
-
 enum class SurveySortType {
     DATE_DESC,
     DATE_ASC,
